+++ crypto/cms/build.info
@@ -2,4 +2,11 @@ LIBS=../../libcrypto
 SOURCE[../../libcrypto]= \
         cms_lib.c cms_asn1.c cms_att.c cms_io.c cms_smime.c cms_err.c \
         cms_sd.c cms_dd.c cms_cd.c cms_env.c cms_enc.c cms_ess.c \
-        cms_pwri.c cms_kari.c
+        cms_pwri.c cms_kari.c cms_rsa.c
+
+IF[{- !$disabled{dh} -}]
+  SOURCE[../../libcrypto]=cms_dh.c
+ENDIF
+IF[{- !$disabled{ed} -}]
+  SOURCE[../../libcrypto]=cms_ecdh.c
+ENDIF
